feat: wired up button to create channel
