+++ apps/mobile/app/[site_id]/(tabs)/home/create-channel.tsx
@@ -1,12 +1,49 @@
-import CreateChannelForm from '@components/features/chat/CreateChannel/CreateChannelForm';
-import { Link, Stack } from 'expo-router';
+import CreateChannelForm, { ChannelCreationForm } from '@components/features/chat/CreateChannel/CreateChannelForm';
+import { Link, router, Stack } from 'expo-router';
 import { Button } from '@components/nativewindui/Button';
 import { Text } from '@components/nativewindui/Text';
 import CrossIcon from '@assets/icons/CrossIcon.svg';
 import { useColorScheme } from '@hooks/useColorScheme';
+import { FormProvider, useForm } from 'react-hook-form';
+import { useFrappeCreateDoc } from "frappe-react-sdk";
+import { ActivityIndicator } from '@components/nativewindui/ActivityIndicator';
 
 export default function CreateChannel() {
+
     const { colors } = useColorScheme()
+    const methods = useForm<ChannelCreationForm>({
+        defaultValues: {
+            type: 'Public',
+            channel_name: '',
+            channel_description: ''
+        }
+    })
+
+    const { handleSubmit, reset: resetForm } = methods
+
+    const { createDoc, error, loading: creatingChannel, reset: resetCreateHook } = useFrappeCreateDoc()
+
+    const reset = () => {
+        resetCreateHook()
+        resetForm()
+    }
+
+    const onSubmit = async (data: ChannelCreationForm) => {
+        return createDoc('Raven Channel', data)
+            .then(result => {
+                if (result) {
+                    // TODO: Show success toast
+                    console.log("Channel created", result)
+                    // TODO: Navigate to channel
+                    // router.push(`/home/${result.name}`)
+                    reset()
+                    resetForm()
+                }
+            }).catch(err => {
+                console.log(err)
+            })
+    }
+
     return <>
         <Stack.Screen options={{
             title: 'Add channel',
@@ -21,14 +58,18 @@ export default function CreateChannel() {
             },
             headerRight() {
                 return (
-                    <Link asChild href="../" relativeToDirectory>
-                        <Button variant="plain" className="ios:px-0">
-                            <Text className="text-primary">Add</Text>
-                        </Button>
-                    </Link>
+                    <Button variant="plain" className="ios:px-0"
+                        onPress={handleSubmit(onSubmit)}
+                        disabled={creatingChannel}>
+                        {creatingChannel ?
+                            <ActivityIndicator size="small" color={colors.primary} /> :
+                            <Text className="text-primary">Add</Text>}
+                    </Button>
                 )
             },
         }} />
-        <CreateChannelForm />
+        <FormProvider {...methods}>
+            <CreateChannelForm />
+        </FormProvider>
     </>
 }
